register fp16 primitive functions
