+++ python/hidet/ir/primitives/cpu/math/__init__.py
@@ -14,3 +14,4 @@ from . import float64
 from . import int32
 from . import int64
 from . import bfloat16
+from . import float16
